requirements.txt
playwright
beautifulsoup4
python-dotenv
sqlalchemy
psycopg2-binary
pillow
instagrapi
beautifulsoup4